@@ -1,108 +1,30 @@
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import {ListUserComponent} from "./list-user/list-user.component";
-import {ListItemsComponent} from "./list-items/list-items.component";
-import {SecurityGuard} from "./security/security.guard";
-import {AuthenticationRoutes} from "./security/authentication/authentication.routing";
-import {ListLogComponent} from "./list-log/list-log.component";
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
+import {authenticationRoute} from "./architecture/authentication/authentication-routing.module";
 import {AboutComponent} from "./about/about.component";
+import {itemRoutes} from "./feature-item/item-routing.module";
+import {userRoutes} from "./feature-user/user-routing.module";
+import {postRoutes} from "./feature-post/post-routing.module";
 
 export const routes: Routes = [
   {
     path: '',
-    redirectTo: 'acesso/login',
-    pathMatch: 'full'
+    children: [
+      ...itemRoutes,
+      ...userRoutes,
+      ...postRoutes
+    ]
   },
   {
-    path: 'items',
-    component: ListItemsComponent,
-    canActivate: [SecurityGuard],
-    data: {
-      security: {
-        roles: [
-          'ROLE_USER_CREATE',
-          'ROLE_USER_READ',
-          'ROLE_USER_UPDATE',
-          'ROLE_USER_DELETE',
-          'ROLE_USER_LISTALL',
-          'ROLE_ITEM_CREATE',
-          'ROLE_ITEM_READ',
-          'ROLE_ITEM_UPDATE',
-          'ROLE_ITEM_DELETE',
-          'ROLE_ITEM_LISTALL'
-        ]
-      }
-    }
-  },
-  {
-    path: 'user',
-    component: ListUserComponent,
-    canActivate: [SecurityGuard],
-    data: {
-      security: {
-        roles: [
-          'ROLE_USER_CREATE',
-          'ROLE_USER_READ',
-          'ROLE_USER_UPDATE',
-          'ROLE_USER_DELETE',
-          'ROLE_USER_LISTALL',
-          'ROLE_ITEM_CREATE',
-          'ROLE_ITEM_READ',
-          'ROLE_ITEM_UPDATE',
-          'ROLE_ITEM_DELETE',
-          'ROLE_ITEM_LISTALL'
-        ]
-      }
-    }
-  },
-  {
-    path: 'log',
-    component: ListLogComponent,
-    canActivate: [SecurityGuard],
-    data: {
-      security: {
-        roles: [
-          'ROLE_USER_CREATE',
-          'ROLE_USER_READ',
-          'ROLE_USER_UPDATE',
-          'ROLE_USER_DELETE',
-          'ROLE_USER_LISTALL',
-          'ROLE_ITEM_CREATE',
-          'ROLE_ITEM_READ',
-          'ROLE_ITEM_UPDATE',
-          'ROLE_ITEM_DELETE',
-          'ROLE_ITEM_LISTALL'
-        ]
-      }
-    }
+    path: "auth",
+    children: [
+      ...authenticationRoute
+    ]
   },
   {
     path: 'about',
-    component: AboutComponent,
-    canActivate: [SecurityGuard],
-    data: {
-      security: {
-        roles: [
-          'ROLE_USER_CREATE',
-          'ROLE_USER_READ',
-          'ROLE_USER_UPDATE',
-          'ROLE_USER_DELETE',
-          'ROLE_USER_LISTALL',
-          'ROLE_ITEM_CREATE',
-          'ROLE_ITEM_READ',
-          'ROLE_ITEM_UPDATE',
-          'ROLE_ITEM_DELETE',
-          'ROLE_ITEM_LISTALL'
-        ]
-      }
-    }
+    component: AboutComponent
   },
-  {
-    path: 'acesso',
-    children: [
-      ...AuthenticationRoutes
-    ]
-  }
 ];
 
 @NgModule({
